zapsentry: Handle the error field properly
To use the right error object to create a Sentry exception, we need to
check for "error" field since that one is created by calling
zap.Error(). In case the field is present, the error object stored under
that key is used when calling NewException.

In case there is no "error" field present, no exception is created. So
creating exceptions is no longer based on log levels, but rather on the
"error" field being present or not.